backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
boto3==1.34.0
botocore==1.34.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.8
google-generativeai==0.3.2
litellm==1.13.0
scrapy==2.11.0
selenium==4.15.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
plotly==5.17.0
streamlit==1.28.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
ffmpeg-python==0.2.0
moviepy==1.0.3